# backend/docs/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.1.0
  title: HomiesHomes Real Estate API
  description: A RESTful API for HomiesHomes real estate listings platform. This API provides endpoints for managing users, properties, bookings, favorites, and agent requests.
  contact:
    name: Mandeep Pandya
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://gammacairo-deltareward-9000.codio-box.uk"
    description: Codio API server
  - url: "https://gammacairo-deltareward-3000.codio-box.uk"
    description: Codio frontend server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Authentication. A valid JWT token must be passed in the Authorization header.
        Example: `Authorization: Bearer {token}`

        You can obtain a token by using the /login endpoint with valid credentials.
        Tokens are valid for 1 hour after issuance.
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed"
paths:
...
  /agent-requests:
    get:
      tags:
        - Agent Requests
      summary: Get all agent requests
      description: Returns a list of all agent role requests (admin only).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of agent requests
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    request_id:
                      type: integer
                      example: 1
                    user_id:
                      type: integer
                      example: 2
                    request_date:
                      type: string
                      format: date-time
                      example: "2023-03-30T12:00:00Z"
                    response_date:
                      type: string
                      format: date-time
                      example: null
                    status:
                      type: string
                      enum: [pending, approved, rejected]
                      example: "pending"
                    request_reason:
                      type: string
                      example: "I'm a professional real estate agent with 5 years of experience."
                    username:
                      type: string
                      example: "john_doe"
                    user_email:
                      type: string
                      example: "john.doe@example.com"
                    user_firstName:
                      type: string
                      example: "John"
                    user_lastName:
                      type: string
                      example: "Doe"
        "401":
          description: Unauthorized - Only admin users can list all agent requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Agent Requests
      summary: Create an agent request
      description: Creates a new agent role request.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - request_reason
              properties:
                request_reason:
                  type: string
                  example: "I'm a professional real estate agent with 5 years of experience."
      responses:
        "201":
          description: Agent request successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Agent request submitted successfully"
                  request:
                    type: object
                    properties:
                      request_id:
                        type: integer
                        example: 1
                      user_id:
                        type: integer
                        example: 2
                      status:
                        type: string
                        enum: [pending, approved, rejected]
                        example: "pending"
                      request_reason:
                        type: string
                        example: "I'm a professional real estate agent with 5 years of experience."
                      request_date:
                        type: string
                        format: date-time
                        example: "2023-03-30T12:00:00Z"
        "400":
          description: Bad request - validation error or user already has a pending request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "You already have a pending agent request"
  /agent-requests/status:
    get:
      tags:
        - Agent Requests
      summary: Get user's agent request status
      description: Returns the status of the current user's agent request.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User's agent request status
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: integer
                    example: 1
                  user_id:
                    type: integer
                    example: 2
                  status:
                    type: string
                    enum: [pending, approved, rejected]
                    example: "pending"
                  request_date:
                    type: string
                    format: date-time
                    example: "2023-03-30T12:00:00Z"
                  response_date:
                    type: string
                    format: date-time
                    example: null
                  request_reason:
                    type: string
                    example: "I'm a professional real estate agent with 5 years of experience."
        "404":
          description: No agent request found for the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "No agent request found"
...